fix splitting factors
