+++ cse-runner-app/src/main/java/com/farao_community/farao/cse/runner/app/CseData.java
@@ -7,16 +7,13 @@
 
 package com.farao_community.farao.cse.runner.app;
 
-import com.farao_community.farao.commons.EICode;
 import com.farao_community.farao.cse.data.CseReferenceExchanges;
 import com.farao_community.farao.cse.data.ntc.Ntc;
 import com.farao_community.farao.cse.data.ntc2.Ntc2;
 import com.farao_community.farao.cse.runner.api.resource.CseRequest;
 import com.farao_community.farao.cse.runner.app.services.FileImporter;
-import com.powsybl.iidm.network.Country;
 
 import java.util.Map;
-import java.util.TreeMap;
 
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
@@ -39,10 +36,9 @@ public class CseData {
     }
 
     public Map<String, Double> getReducedSplittingFactors() {
-        if (reducedSplittingFactors != null) {
-            return reducedSplittingFactors;
+        if (reducedSplittingFactors == null) {
+            reducedSplittingFactors = getNtc().computeReducedSplittingFactors();
         }
-        reducedSplittingFactors = convertSplittingFactors(getNtc().computeReducedSplittingFactors());
         return reducedSplittingFactors;
     }
 
@@ -104,14 +100,4 @@ public class CseData {
         this.preProcesedNetworkUrl = preProcesedNetworkUrl;
     }
 
-    private Map<String, Double> convertSplittingFactors(Map<String, Double> tSplittingFactors) {
-        Map<String, Double> splittingFactors = new TreeMap<>();
-        tSplittingFactors.forEach((key, value) -> splittingFactors.put(toEic(key), value));
-        splittingFactors.put(toEic("IT"), -splittingFactors.values().stream().reduce(0., Double::sum));
-        return splittingFactors;
-    }
-
-    private String toEic(String country) {
-        return new EICode(Country.valueOf(country)).getAreaCode();
-    }
 }
